Refactor sidebar buttons into a component
