+++ app/views/catalog/_show_sidebar.html.erb
@@ -1,7 +1,7 @@
 <nav class="show_sidebar">
   <%= render ExternalLinksComponent.new(document: @document) %>
   <hr>
-  <%= render 'items/button_link_list', buttons: @buttons_presenter.buttons %>
+  <%= render SidebarControlsComponent.new(solr_document: @document, manager: can?(:manage_item, @cocina)) %>
 </nav>
 
 <%= render 'catalog/show_embargo_sidebar', document: @document %>
