--- app/views/items/_button_link_list.html.erb
@@ -1,5 +1,0 @@
-<div class='btn-group-vertical argo-show-btn-group' role='group' aria-label='Argo action buttons'>
-  <% buttons.each do |button| %>
-    <%= render ActionButton.new(**button) %>
-  <% end %>
-</div>
